Refer to facade root
Referring to a facade class directly defeats the purpose of a facade. This may have been picked up from the Illuminate library which do refer to facades directly, but this is an internal pattern. It may also be to help IDE's resolve facade methods, this is not an issue in October v2 onwards.

Hit one reference here, recommend doing them all